Update to match the jonesforth.c dictionary format.
The dictionary used by the C version differs from the original. This is
mostly driven because the primitives and builtins are named using C
format strings (e.g. '\0' terminated). This means we can discover how long
a name is without reading the flags... that in turn allows us to move the
flags to just before the codeword. This dictionary is slightly less
space efficient but makes it easy to move from the codeword back to the
word name.